+++ IEX_29id/devices/detectors.py
@@ -4,8 +4,6 @@ from epics import caget, caput
 from IEX_29id.scans.setup import Scan_FillIn,  Scan_Go
 from math import *
 
-import numpy as np
-import numpy.polynomial.polynomial as poly
 #import pandas as pd
 
 # def Kappa_Detector_Offset(detector):
