+++ IEX_29id/devices/undulator.py
@@ -2,7 +2,7 @@ from IEX_29id.devices.energy import Switch_IDMode, Open_MainShutter, Check_Grati
 from epics import caput,caget, poly
 from time import sleep
 from IEX_29id.utils.exp import WaitForPermission, Check_MainShutter
-from IEX_29id.utils.misc import dateandtime, RangeUp
+from IEX_29id.utils.misc import dateandtime
 
 
 def ID_Calc(grt,mode,hv):    # Mode = state (0=RCP,1=LCP,2=V,3=H)
